CTFPatch import
RESP AND BUGFIX ARE NOT COMPILING CORRECTLY. RESP is missing SetDeadSpectator, a very real Zandronum-only function, and BUGFIX is missing round - it compiles just fine when it's done in Slade with the appropriate prerequisites, but doesn't compile correctly with the doom archive manager files.

The compiled files in /acs are, however, accurate. You shouldn't need to compile RESP again and any changes to BUGFIX will need to be made in Slade. I'll be home to fix this in a few hours.