Actual fix to Knightman sky horizons
this is probably the first instance I've had so far of beginning and ending the process of bugfixing with absolutely no idea whatsoever of what actually caused the bug. there's no way the issue with KnightmanCrusherNoRip was that it used goto SpawnPOOP instead of loop (to SpawnPOOP), but that seems to have fixed it in the only environment I can test this in.

Hell.